jenkins-shared-library

+++ vars/nodejsEKS.groovy
@@ -15,6 +15,7 @@ def call(Map configMap) {
             account_id = pipelineGlobals.account_id()
             component  = configMap.get('component')
             project  =   configMap.get('project')
+            def releaseExists = ""
         }
         stages {
             stage('read the version') {
@@ -57,13 +58,27 @@ def call(Map configMap) {
             }
             stage('Deploy') {
                 steps {
-                    sh """
-                    aws eks update-kubeconfig --region ${region} --name ${project}-dev
-                    cd helm
-                    sed -i 's/IMAGE_VERSION/${appVersion}/g' values.yaml
-                    helm install ${component} -n ${project} .
-
-                """
+                    script {
+                        releaseExists = sh(script: "helm list -A --short | grep -w ${component} || true", retrunStdout: true).trim()
+                        if(releaseExists.isEmpty()){
+                            echo "${component} not installed yet, first installation"
+                            sh """
+                                aws eks update-kubeconfig --region ${region} --name ${project}-dev
+                                cd helm
+                                sed -i 's/IMAGE_VERSION/${appVersion}/g' values.yaml
+                                helm install ${component} -n ${project} .
+                            """ 
+                        }
+                        else{
+                            echo "${component} exists, running upgrade"
+                            sh """
+                                aws eks update-kubeconfig --region ${region} --name ${project}-dev
+                                cd helm
+                                sed -i 's/IMAGE_VERSION/${appVersion}/g' values.yaml
+                                helm upgrade ${component} -n ${project} .
+                            """    
+                        }
+                    }
                 }
             }
         // stage ('nexus artifact upload') {
